backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
pydantic
pytest
pytest-asyncio
httpx
